+++ extension.js
@@ -39,6 +39,8 @@ let compatmode = false;
  */
 async function activate(context) {
 
+	const defaultWorkspaceConfiguration = JSON.stringify({ folders: [{ path: "." }] });
+
 	compatmode = vscode.workspace.getConfiguration('figura').get('documentation.enableDocumentation');
 
 	if (compatmode) {
@@ -106,11 +108,15 @@ async function activate(context) {
 				let items = [];
 				const linePrefix = document.lineAt(position).text.substr(0, position.character);
 
-				if (linePrefix.match(/\.\w+$/)) return [];
+				if (linePrefix.match(/[\.:]\w+$/)) return [];
 
 				for (let i = 0; i < rootgroups.length; i++) {
 					let item = new vscode.CompletionItem(rootgroups[i].group.words[0], rootgroups[i].group.type);
-					item.commitCharacters = ['.'];
+					if (rootgroups[i].group.documentations != undefined) {
+						item.detail = 'Documentation available ->';
+						item.documentation = rootgroups[i].group.documentations[0];
+					}
+					item.commitCharacters = ['.', ':'];
 					if (!compatmode || rootgroups[i].ignoreCompat)
 						items.push(item);
 				}
@@ -130,13 +136,14 @@ async function activate(context) {
 
 				for (let i = 0; i < rootgroups.length; i++) {
 					if (!compatmode || rootgroups[i].ignoreCompat)
-						items = items.concat(browse(document, position, linePrefix, rootgroups[i].group.subgroups, '\\b' + rootgroups[i].group.words[0] + '\\.'));
+						items = items.concat(browse(document, position, linePrefix, rootgroups[i].group.subgroups, '\\b' + rootgroups[i].group.words[0] + '[\\.:]'));
 				}
 
 				return items;
 			}
 		},
-		'.'
+		'.',
+		':'
 	);
 
 	// recursively find matching words
@@ -149,23 +156,33 @@ async function activate(context) {
 			const group = groups[i];
 			for (let n = 0; n < group.words.length; n++) {
 				let hasSpaces = group.words[n].includes(' ');
+				let isSelfCall = group.selfcalls != undefined ? group.selfcalls[n] : false;
 				if (linePrefix.match(new RegExp(pattern + '$')) && group.showSuggestion) {
 					if (compatmode && group.type == vscode.CompletionItemKind.Method) continue; // in compat mode dont show methods, even when ignoreCompat is true
 
 					let item = new vscode.CompletionItem(group.words[n], group.type);
+					if (group.documentations != undefined && group.documentations[n] != undefined) {
+						item.detail = 'Documentation available ->';
+						item.documentation = group.documentations[n];
+					}
 					if (hasSpaces) {
 						item.insertText = '["' + group.words[n] + '"]';
 						let range = new vscode.Range(new vscode.Position(position.line, position.character - 1), position);
-						item.additionalTextEdits = [new vscode.TextEdit(range, '')]; // remove the '.'
+						item.additionalTextEdits = [new vscode.TextEdit(range, '')]; // remove the '.' or ':'
+					}
+					if (isSelfCall) {
+						item.insertText = ':' + group.words[n];
+						let range = new vscode.Range(new vscode.Position(position.line, position.character - 1), position);
+						item.additionalTextEdits = [new vscode.TextEdit(range, '')]; // remove the '.' or ':'
 					}
 					if (group.type == vscode.CompletionItemKind.Method) item.commitCharacters = ['('];
-					else if (group.type == vscode.CompletionItemKind.Property) item.commitCharacters = ['.'];
+					else if (group.type == vscode.CompletionItemKind.Property) item.commitCharacters = ['.', ':'];
 
 					items.push(item);
 				}
 				let nextpattern;
-				if (hasSpaces) nextpattern = pattern.substring(0, pattern.length - 2) + '\\[\\"' + group.words[n] + '\\"\\]' + '\\.'; // in this case remove the \. from the pattern and use ["word"]
-				else nextpattern = pattern + group.words[n] + '\\.';
+				if (hasSpaces) nextpattern = pattern.substring(0, pattern.length - 2) + '\\[\\"' + group.words[n] + '\\"\\]' + '[\\.:]'; // in this case remove the \. from the pattern and use ["word"]
+				else nextpattern = pattern + group.words[n] + '[\\.:]';
 				items = items.concat(
 					browse(document, position, linePrefix, group.subgroups, nextpattern)
 				);