+++ src/wordgroup.js
@@ -1,7 +1,9 @@
 class WordGroup {
-    constructor(words, type, showSuggestion) {
+    constructor(words, type, showSuggestion, documentations, selfcalls) {
         this.words = words
         this.type = type
+        this.documentations = documentations // array, same length as words, defines documentation for each word
+        this.selfcalls = selfcalls // array, same length as words, defines if that word should use : instead of .
         if (showSuggestion == undefined) {
             this.showSuggestion = true
         }